refactor: 패키지 이동 및 테스트 코드 수정
